# Improved generate_event_sed