`generate_event_ses` was calling `.get_background_sids` inside the SES loop, i.e. 9999 more times than needed for 10,000 SES. Having extracted `.get_background_sids` from the loop `sampling ruptures` is a bit faster but not much (i.e. 30% faster, but we would need to it to be 522x faster). For the time being, we should rely on time-independence and use a single SES of 10,000 years. Here are some numbers:

```
master, 10000 SES x 1 year
=========================== ========= ========= =======
operation                   time_sec  memory_mb counts
=========================== ========= ========= =======
total compute_losses        795       12        24
sampling ruptures           780       0.0       240,000
complete runtime            99        12        1

master, 1 SES x 10,000 years
=========================== ========= ========= ======
operation                   time_sec  memory_mb counts
=========================== ========= ========= ======
total compute_losses        11        12        24
complete runtime            1.797     10        1
sampling ruptures           1.495     0.0       24

this branch, 10,000 SES x 1 year
=========================== ========= ========= =======
operation                   time_sec  memory_mb counts
=========================== ========= ========= =======
total compute_losses        575       12        24
sampling ruptures           560       0.0       240,000
complete runtime            71        14        1
```